EdgeModules/modules/avroFileWriter/requirements.txt
azure-iothub-device-client~=1.4.3
fastavro
pyyaml